fix(governance): satisfy the repo gates this change legitimately trips
Module Size Guard: the WSID ledger write pushed build-reviewers.ts to 863 LOC,
over the 800 ceiling. Rather than baseline a new oversized file, extracted the
whole architecture-advisory concern -- advisory shaping, reference-doc
promotion, and the ledger write -- into lib/integrate/architecture-advisory-ledger.ts
(758 LOC left behind). That is better cohesion, not gate appeasement: it puts
the advisory pipeline in one module and keeps the two OPPOSITE fail postures in
separate files, so fail-OPEN (advisory) is not read as fail-CLOSED (triage
drain). Re-exported from the lib/build-reviewers shim so callers are unmoved.

Two baselined files grew and are re-baselined with cause:
  agent-grants.ts 831->851 -- three gating entries plus the explanation of why
    a missing entry is a sealed door. This growth IS the BI-88B77204 fix.
  mcp-tools.ts 2007->2008 -- one import, replacing a hardcoded coworker name
    with the constant bound to agent-identity.
The same pass retightens AgentCoworkerPanel.tsx 1266->1253 from a peer's shrink.

Data-Impact Gate: added a manifest for the DecisionInteraction columns. Records
that the backfill is EXACT rather than approximate -- the profession gate had
never executed when the historical rows were written (169 platform, 10
organization, 0 profession at the time of change), so no row's gate is
ambiguous -- and that the change is additive with nothing dropped or
overwritten.

Docs Impact Gate: documented the decision log properly rather than attesting it
away, because the user-visible behaviour genuinely changed. The user guide now
covers reading the three tiers, and two things that would otherwise mislead an
operator: a fallback-marked WSID row means "the role decided, but not yet from
its own craft" (a prompt to grow that corpus, not evidence of health), and a
tier reading "never used" is a question rather than a conclusion -- it can mean
nothing is wired to call that gate, which is exactly what BI-88B77204 was.

Refs BI-1BE30A9A, BI-88B77204, BI-D6CFE63A, EP-0AF96937
