+++ scripts/module-size-baseline.txt
@@ -2,7 +2,7 @@ apps/web/app/api/mcp/v1/route.test.ts	1260
 apps/web/app/api/v1/edge/discovery-runs/route.test.ts	982
 apps/web/components/admin/BusinessModelBuilder.tsx	821
 apps/web/components/admin/McpTokenManager.tsx	1373
-apps/web/components/agent/AgentCoworkerPanel.tsx	1266
+apps/web/components/agent/AgentCoworkerPanel.tsx	1253
 apps/web/components/agent/AgentMessageInput.tsx	1034
 apps/web/components/ea/EaCanvas.tsx	1047
 apps/web/components/inventory/TopologyGraph.tsx	1031
@@ -41,7 +41,7 @@ apps/web/lib/integrate/build-orchestrator.ts	1751
 apps/web/lib/integrate/sandbox/build-branch.ts	852
 apps/web/lib/marketing.ts	1367
 apps/web/lib/mcp-tools-backlog.test.ts	874
-apps/web/lib/mcp-tools.ts	2007
+apps/web/lib/mcp-tools.ts	2008
 apps/web/lib/mcp/packs/backlog-pack.ts	1318
 apps/web/lib/mcp/packs/contribution-hive-pack.ts	858
 apps/web/lib/mcp/packs/sandbox-pack.ts	920
@@ -54,7 +54,7 @@ apps/web/lib/self-upgrade/quiescence.test.ts	865
 apps/web/lib/self-upgrade/quiescence.ts	1469
 apps/web/lib/skills/evidence-search.ts	803
 apps/web/lib/tak/agent-grants.test.ts	903
-apps/web/lib/tak/agent-grants.ts	831
+apps/web/lib/tak/agent-grants.ts	851
 apps/web/lib/tak/agent-routing.ts	965
 apps/web/lib/tak/agentic-loop.test.ts	2311
 apps/web/lib/tak/agentic-loop.ts	2683
